Оптимизированна работа c баззой данных, исправлены  некоторые код-стайл моменты

+++ src/main/java/ru/yandex/practicum/filmorate/controllers/ReviewController.java
@@ -18,8 +18,8 @@ public class ReviewController {
     private final ReviewService reviewService;
 
     @GetMapping
-    public List<Review> getUsefulReviews(@RequestParam(defaultValue = "0") Integer filmId,
-                                         @RequestParam(defaultValue = "10") Integer count) {
+    public List<Review> getUsefulReviews(@RequestParam(defaultValue = "0") Long filmId,
+                                         @RequestParam(defaultValue = "10") Long count) {
         if (filmId == 0) {
             return reviewService.getReviews();
         }
@@ -27,7 +27,7 @@ public class ReviewController {
     }
 
     @GetMapping("/{id}")
-    public Review getReviewById(@PathVariable Integer id) {
+    public Review getReviewById(@PathVariable Long id) {
         return reviewService.getReviewById(id);
     }
 
@@ -42,28 +42,28 @@ public class ReviewController {
     }
 
     @DeleteMapping("/{id}")
-    public Boolean deleteReview(@PathVariable Integer id) {
+    public Boolean deleteReview(@PathVariable Long id) {
         return reviewService.deleteReview(id);
     }
 
 
     @PutMapping("/{id}/like/{userId}")
-    public Review likeReview(@PathVariable Integer id, @PathVariable Integer userId) {
+    public Review likeReview(@PathVariable Long id, @PathVariable Long userId) {
         return reviewService.likeReview(id, userId);
     }
 
     @PutMapping("/{id}/dislike/{userId}")
-    public Review dislikeReview(@PathVariable Integer id, @PathVariable Integer userId) {
+    public Review dislikeReview(@PathVariable Long id, @PathVariable Long userId) {
         return reviewService.dislikeReview(id, userId);
     }
 
     @DeleteMapping("/{id}/like/{userId}")
-    public Review deleteLike(@PathVariable Integer id, @PathVariable Integer userId) {
+    public Review deleteLike(@PathVariable Long id, @PathVariable Long userId) {
         return reviewService.deleteLike(id, userId);
     }
 
     @DeleteMapping("/{id}/dislike/{userId}")
-    public Review deleteDislike(@PathVariable Integer id, @PathVariable Integer userId) {
+    public Review deleteDislike(@PathVariable Long id, @PathVariable Long userId) {
         return reviewService.deleteDislike(id, userId);
     }
